Cleaned up small warnings on i386
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@6406 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/arch/i386/init.c
@@ -49,8 +49,9 @@ mmu_init(void)
     flush_tlb_hard();
 
     /* Load TSS */
-    gdt[(GD_TSS >> 3)] = SEG_BASELO(&tss) | SEG_LIM(sizeof(tss) - 1) |
-			 SEG_P | SEG_A | SEG_TSSA;
+    gdt[(GD_TSS >> 3)] = SEG_BASELO((uintptr_t)&tss) |
+							SEG_LIM(sizeof(tss) - 1) |
+							SEG_P | SEG_A | SEG_TSSA;
     ltr(GD_TSS);
 }
